extracting out core model classes for salesforce and config
this is to support creating a sttp based salesforce api client
