+++ handlers/sf-gocardless-sync/src/test/scala/com/gu/sf_gocardless_sync/salesforce/SalesforceDDMandateEffectsTest.scala
@@ -1,8 +1,9 @@
 package com.gu.sf_gocardless_sync.salesforce
 
 import com.gu.effects.{GetFromS3, RawEffects}
-import com.gu.salesforce.SalesforceAuthenticate.SFAuthConfig
+import com.gu.salesforce.SFAuthConfig
 import com.gu.salesforce.SalesforceClient
+import com.gu.salesforce.SalesforceReads._
 import com.gu.sf_gocardless_sync.SyncSharedObjects.{BankAccountNumberEnding, BankName, GoCardlessMandateID, MandateCreatedAt, Reference}
 import com.gu.sf_gocardless_sync.salesforce.SalesforceDDMandate.Create.WireNewMandate
 import com.gu.sf_gocardless_sync.salesforce.SalesforceDDMandate.Update.WirePatchMandate
