+++ lib/salesforce/core/src/main/scala/com/gu/salesforce/SalesforceAuth.scala
@@ -1,0 +1,4 @@
+package com.gu.salesforce
+
+// the WireResponse model is the same as the domain model, so keep a friendly name
+case class SalesforceAuth(access_token: String, instance_url: String)
