+++ lib/salesforce/effects/src/main/scala/com/gu/salesforce/SalesforceAuthenticate.scala
@@ -1,45 +1,15 @@
 package com.gu.salesforce
 
 import com.gu.util.Logging
-import com.gu.util.config.ConfigLocation
 import com.gu.util.reader.Types._
 import com.gu.util.resthttp.RestOp._
 import com.gu.util.resthttp.RestRequestMaker.toClientFailableOp
 import com.gu.util.resthttp.{HttpOp, LazyClientFailableOp}
 import okhttp3.{FormBody, Request, Response}
-import play.api.libs.json.{Json, Reads}
+import play.api.libs.json.Json
+import com.gu.salesforce.SalesforceReads._
 
 object SalesforceAuthenticate extends Logging {
-
-  case class SFAuthConfig(
-    url: String,
-    client_id: String,
-    client_secret: String,
-    username: String,
-    password: String,
-    token: String
-  )
-
-  object SFAuthConfig {
-    implicit val reads: Reads[SFAuthConfig] = Json.reads[SFAuthConfig]
-    implicit val location = ConfigLocation[SFAuthConfig](path = "sfAuth", version = 1)
-  }
-
-  object SFAuthTestConfig {
-    implicit val reads: Reads[SFAuthConfig] = SFAuthConfig.reads
-    implicit val location = ConfigLocation[SFAuthConfig](path = "TEST/sfAuth", version = 1)
-  }
-
-  object SFExportAuthConfig {
-    val location = ConfigLocation[SFAuthConfig](path = "sfExportAuth", version = 1)
-  }
-
-  // the WireResponse model is the same as the domain model, so keep a friendly name
-  case class SalesforceAuth(access_token: String, instance_url: String)
-
-  object SalesforceAuth {
-    implicit val salesforceAuthReads: Reads[SalesforceAuth] = Json.reads[SalesforceAuth]
-  }
 
   private def buildAuthRequest(config: SFAuthConfig) = {
     import config._
