+++ lib/salesforce/effects/src/main/scala/com/gu/salesforce/SalesforceConstants.scala
