+++ lib/salesforce/effects/src/main/scala/com/gu/salesforce/TypesForSFEffectsData.scala
