+++ lib/salesforce/effects/src/test/scala/com/gu/salesforce/dev/SFEffectsData.scala
@@ -11,5 +11,4 @@ object SFEffectsData {
 
   // this contact we can update the identity id freely
   val updateIdentityIdEmailAndFirstNameContact = SFContactId("0036E00000Ho05i")
-
 }
